pyproject.toml
[project]
name = "amplifier-app-api"
version = "0.2.0"
description = "REST API service for Amplifier AI development platform"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Organization" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.3",
    "httpx>=0.28.1",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "amplifier-core",
    "amplifier-foundation",
    "opencensus-ext-azure>=1.1.13",
]

[project.scripts]
amplifier-service = "amplifier_app_api.main:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["amplifier_app_api"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

# Using specific commit hashes to prevent breaking changes
[tool.uv.sources]
amplifier-core = { git = "https://github.com/microsoft/amplifier-core.git", rev = "976fb87335ffe398cf0c1bd1bcd3c2f2c154fc3c" }
amplifier-foundation = { git = "https://github.com/microsoft/amplifier-foundation.git", rev = "412fcb51980523c77339c32ef1ba3bc80c13b680" }

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--import-mode=importlib"
asyncio_mode = "strict"
markers = [
    "e2e: End-to-end tests with real HTTP server (slow)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)
